Move npm to it's own plugin
